Use /bin/bash to run the shell scripts.
Debian changed /bin/sh from being a link to bash to being a
link to dash. One consequence is that it broke scripts that used

  #!/bin/sh

to run with a basic shell, but then used extended bash functionality.

The problem is fixed by having scripts that depend on bash extra
functionality call #!/bin/bash (as they should).

For a discussion about the topic, see for example:
https://askubuntu.com/questions/976485/