Upgrade to tomcat 9 with jre11.
This commit converts the base docker image in `core` to tomcat:9-jre11-temurin
which satisfies all the requirements we need to run the latest dhis2. See:
https://docs.dhis2.org/en/manage/performing-system-administration/dhis-core-version-238/installation.html

(The only difference is that instead of using the full JDK, we just use the JRE,
which is really the only thing needed.)

The image tomcat:9-jre11-temurin is based on ubuntu, not alpine anymore. This
is the main source of the changes described below.

We make sure to call bash as the shell to run the different scripts (some of
which need the shell to be bash indeed and assumed wrongly that sh would always
be bash).

We changed
  config/dhis2-core-entrypoint.sh
  images/dhis2-core/docker-entrypoint.sh

to make them both basically identical, and based on the current latest tagged
version of dhis2-core entrypoint:
https://github.com/dhis2/dhis2-core/blob/2.36.11.1/docker/tomcat-debian/docker-entrypoint.sh

Instead of
  su-exec tomcat
which is a command found in alpine linux, we use the equivalent
  setpriv --reuid=tomcat --regid=tomcat --init-groups

We also remove the need to keep different apk packages in the repository, since
we install all the needed with `apt-get install` in the dhis2-core Dockerfile.