Backwards compatibility when running with old cores.
Old core images use alpine linux. That means that the commands executed
on their entrypoint (docker-entrypoint.sh) can be "su-exec" (which is
what alpine linux uses to run something as a different user), but not
"setpriv" (which is what ubuntu and others use for the same thing).

This change tests which kind of system the core is, and runs the
appropriate command on entrypoint.